Move pages back to pages root folder
