--- src/pages/404.js
@@ -1,1 +1,0 @@
-export { default } from '@/pages/404';
