+++ src/pages/500.page.js
