--- src/pages/[...params].js
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from '@/pages/[...params]';
