--- src/pages/_app.js
@@ -1,1 +1,0 @@
-export { default } from '@/pages/_app';
