--- src/pages/_document.js
@@ -1,1 +1,0 @@
-export { default } from '@/pages/_document';
