+++ src/pages/copyright.page.tsx
