--- src/pages/copyright/index.tsx
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from '@/components/pages/copyright';
