--- src/pages/create.ts
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from '@/pages/create';
