--- src/pages/dashboard.tsx
@@ -1,10 +1,0 @@
-import Dashboard, {
-  getServerSideProps as dashboardGetServerSideProps,
-} from '@/pages/dashboard';
-
-const DashboardWrapper = (props) => {
-  return <Dashboard {...props} />;
-};
-
-export default DashboardWrapper;
-export const getServerSideProps = dashboardGetServerSideProps;
