--- src/pages/events/[eventId]/availability.js
@@ -1,4 +1,0 @@
-export {
-  default,
-  getServerSideProps,
-} from '@/pages/events/[eventId]/availability';
