--- src/pages/login/[language].js
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from '@/pages/login/[language]';
