+++ src/pages/manage/movies/PictureDeleteModal.tsx
