+++ src/pages/manage/movies/PictureUploadModal.tsx
