--- src/pages/manage/movies/[eventId]/edit.tsx
@@ -1,4 +1,0 @@
-export {
-  default,
-  getServerSideProps,
-} from '@/pages/manage/movies/[eventId]/edit';
