--- src/pages/manage/movies/create.tsx
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from '@/pages/manage/movies/create';
