--- src/pages/manage/productions/create.js
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from '@/pages/manage/productions/create';
