--- src/pages/manage/productions/index.js
@@ -1,1 +1,0 @@
-export { default, getServerSideProps } from '@/pages/manage/productions/index';
