--- src/pages/places/[placeId]/availability.js
@@ -1,4 +1,0 @@
-export {
-  default,
-  getServerSideProps,
-} from '@/pages/places/[placeId]/availability';
